refactor: add deployable object for files and apps
