+++ inc/file.class.php
@@ -36,20 +36,12 @@ if (!defined('GLPI_ROOT')) {
 /**
  * @since 0.1.30
  */
-class PluginFlyvemdmFile extends CommonDBTM {
+class PluginFlyvemdmFile extends PluginFlyvemdmDeployable {
 
-   // name of the right in DB
+   /**
+    * @var string $rightname name of the right in DB
+    */
    static $rightname = 'flyvemdm:file';
-
-   /**
-    * @var bool $usenotepad enable notepad for the itemtype (GLPi < 0.85)
-    */
-   protected $usenotepad = true;
-
-   /**
-    * @var bool $usenotepad enable notepad for the itemtype (GLPi >=0.85)
-    */
-   protected $usenotepadRights = true;
 
    /**
     * Localized name of the type
@@ -194,31 +186,6 @@ class PluginFlyvemdmFile extends CommonDBTM {
    }
 
    /**
-    * get the URL to download the file
-    * @return string|boolean URL of the file
-    */
-   public function getFileURL() {
-      $config = Config::getConfigurationValues('flyvemdm', ['deploy_base_url']);
-      $deployBaseURL = $config['deploy_base_url'];
-      if ($deployBaseURL === null) {
-         return false;
-      }
-
-      $URL = $deployBaseURL . '/file/' . $this->fields['source'];
-      return $URL;
-   }
-
-   /**
-    * Create a directory
-    * @param string $dir
-    */
-   protected function createEntityDirectory($dir) {
-      if (!is_dir($dir)) {
-         @mkdir($dir, 0770, true);
-      }
-   }
-
-   /**
     * @return array
     */
    public function getSearchOptionsNew() {
@@ -268,10 +235,6 @@ class PluginFlyvemdmFile extends CommonDBTM {
       ]);
    }
 
-   public function post_addItem() {
-      global $DB;
-   }
-
    /**
     * Actions done after the getFromFB method
     */
@@ -285,7 +248,8 @@ class PluginFlyvemdmFile extends CommonDBTM {
          if (isAPI()
             && (isset($_SERVER['HTTP_ACCEPT']) && $_SERVER['HTTP_ACCEPT'] == 'application/octet-stream'
                || isset($_GET['alt']) && $_GET['alt'] == 'media')) {
-            $this->sendFile(); // and terminate script
+            $this->sendFile(FLYVEMDM_FILE_PATH . "/" . $this->fields['source'],
+               $this->fields['name'], $this->fields['filesize']); // and terminate script
          }
       }
    }
@@ -299,119 +263,13 @@ class PluginFlyvemdmFile extends CommonDBTM {
       $itemtype = $this->getType();
       $itemId = $this->getID();
 
-      $task = new PluginFlyvemdmTask();
-      $tasks = $task->find("`itemtype`='$itemtype' AND `items_id`='$itemId'");
-      foreach ($tasks as $taskId => $taskRow) {
-         $notifiableType = $taskRow['itemtype_applied'];
-         $notifiable = new $notifiableType();
-         $notifiableId = $taskRow['items_id_applied'];
-         if ($notifiable->getFromDB($notifiableId)) {
-            Toolbox::logInFile('php-errors',
-               "Plugin Flyvemdm : Could not find notifiable id = '$notifiableId'");
-            continue;
-         }
-         if ($task->getFromDB($taskId)) {
-            $task->publishPolicy($notifiable);
-            $task->createTaskStatuses($notifiable);
-         }
-      }
+      $this->publishTaskAction($itemtype, $itemId, new PluginFlyvemdmTask);
    }
 
    public function post_purgeItem() {
-      $filename = FLYVEMDM_FILE_PATH . "/" . $this->fields['source'];
-      if (is_writeable($filename)) {
-         unlink($filename);
-      }
+      $this->unlinkLocalFile(FLYVEMDM_FILE_PATH . "/" . $this->fields['source']);
    }
 
-   /**
-    * Sends a file
-    */
-   protected function sendFile() {
-      $streamSource = FLYVEMDM_FILE_PATH . "/" . $this->fields['source'];
-
-      // Ensure the file exists
-      if (!file_exists($streamSource) || !is_file($streamSource)) {
-         header("HTTP/1.0 404 Not Found");
-         exit(0);
-      }
-
-      // Download range defaults to the full file
-      // get file metadata
-      $size = $this->fields['filesize'];
-      $begin = 0;
-      $end = $size - 1;
-      $mimeType = 'application/octet-stream';
-      $time = date('r', filemtime($streamSource));
-
-      // Open the file
-      $fileHandle = @fopen($streamSource, 'rb');
-      if (!$fileHandle) {
-         header("HTTP/1.0 500 Internal Server Error");
-         exit(0);
-      }
-
-      // set range if specified by the client
-      if (isset($_SERVER['HTTP_RANGE'])) {
-         $matches = null;
-         if (preg_match('/bytes=\h*(\d+)?-(\d*)[\D.*]?/i', $_SERVER['HTTP_RANGE'], $matches)) {
-            if (!empty($matches[1])) {
-               $begin = intval($matches[1]);
-            }
-            if (!empty($matches[2])) {
-               $end = min(intval($matches[2]), $end);
-            }
-         }
-      }
-
-      // seek to the begining of the range
-      $currentPosition = $begin;
-      if (fseek($fileHandle, $begin, SEEK_SET) < 0) {
-         header("HTTP/1.0 500 Internal Server Error");
-         exit(0);
-      }
-
-      // send headers to ensure the client is able to detect a corrupted download
-      // example : less bytes than the expected range
-      // send meta data
-      // setup client's cache behavior
-      header("Expires: Mon, 26 Nov 1962 00:00:00 GMT");
-      header('Pragma: private'); /// IE BUG + SSL
-      header('Cache-control: private, must-revalidate'); /// IE BUG + SSL
-      header("Content-disposition: attachment; filename=\"" . $this->fields['name'] . "\"");
-      header("Content-type: $mimeType");
-      header("Last-Modified: $time");
-      header('Accept-Ranges: bytes');
-      header('Content-Length:' . ($end - $begin + 1));
-      header("Content-Range: bytes $begin-$end/$size");
-      header("Content-Transfer-Encoding: binary\n");
-      header('Connection: close');
-
-      // Prepare HTTP response
-      if ($begin > 0 || $end < $size - 1) {
-         header('HTTP/1.0 206 Partial Content');
-      } else {
-         header('HTTP/1.0 200 OK');
-      }
-
-      // Sends bytes until the end of the range or connection closed
-      while (!feof($fileHandle) && $currentPosition < $end && (connection_status() == 0)) {
-         // allow a few seconds to send a few KB.
-         set_time_limit(10);
-         $content = fread($fileHandle, min(1024 * 16, $end - $currentPosition + 1));
-         if ($content === false) {
-            header("HTTP/1.0 500 Internal Server Error", true); // Replace previously sent headers
-            exit(0);
-         } else {
-            print $content;
-         }
-         flush();
-         $currentPosition += 1024 * 16;
-      }
-
-      // Endnow to prevent any unwanted bytes
-      exit(0);
-   }
 
    /**
     * Deletes files related to the entity being purged
